+++ source/include/customtime.h
@@ -49,16 +49,16 @@ namespace CustomTime {
     void resetToDefaults();
     // This function is called frequently from other functions, ensuring that we check and sync time if needed
     bool isTimeSynched();
-    void setOffset(int gmtOffset, int dstOffset);
+    void setOffset(int32_t gmtOffset, int32_t dstOffset);
 
-    unsigned long long getUnixTime();
-    unsigned long long getUnixTimeMilliseconds();
+    uint64_t getUnixTime();
+    uint64_t getUnixTimeMilliseconds();
     void getTimestamp(char* buffer, size_t bufferSize);
     void getTimestampIso(char* buffer, size_t bufferSize);
     void getDate(char* buffer, size_t bufferSize);
     void getDateIso(char* buffer, size_t bufferSize);
 
-    unsigned long long getMillisecondsUntilNextHour();
+    uint64_t getMillisecondsUntilNextHour();
 
     void timestampFromUnix(time_t unix, char* buffer, size_t bufferSize);
     void timestampIsoFromUnix(time_t unix, char* buffer, size_t bufferSize);
