+++ source/include/mqtt.h
@@ -110,9 +110,9 @@
 
 struct DebugFlagsRtc {
     bool enableMqttDebugLogging;
-    unsigned long long mqttDebugLoggingDurationMillis;
-    unsigned long long mqttDebugLoggingEndTimeMillis;
-    unsigned int signature;
+    uint64_t mqttDebugLoggingDurationMillis;
+    uint64_t mqttDebugLoggingEndTimeMillis;
+    uint32_t signature;
     // Since this struct will be used in an RTC_NOINIT_ATTR, we cannot initialize it in the constructor
 };
 
@@ -149,6 +149,6 @@ namespace Mqtt
     void requestStatisticsPublish();
     
     // Public methods for pushing data to queues
-    void pushLog(const char* timestamp, unsigned long long millisEsp, const char* level, unsigned int coreId, const char* function, const char* message);
+    void pushLog(const char* timestamp, uint64_t millisEsp, const char* level, uint32_t coreId, const char* function, const char* message);
     void pushMeter(const PayloadMeter& payload);
 }
